+++ gDRplots/R/plot-combo-heatmaps.R
@@ -21,7 +21,7 @@
 #' @param colors_vec_excess character vector of colors (valid name or hex codes) used in the heatmap
 #'    for excess values; the default is the blue-light grey-red color scale
 #' @param limit numeric vector of length two providing limits of the scale. 
-#'    Use NA to refer to the existing minimum or maximumdescription
+#'    Use NA to refer to the existing minimum or maximum
 #' @param no_breaks numeric number of breaks on scale
 #' @param swap_axes logical flag indicating whether to swap the axes with drugs of the heatmap
 #' @param show_values logical flag indicating whether to show values of the metric on the heatmap
@@ -162,7 +162,7 @@ heatmap_combo_metrics <- function(
   } else {
     "IC"
   }
-  legend_lbl_iso <- paste0(label_prefix, 100 - 100 * as.numeric(available_iso_lvl))
+  legend_lbl_iso <- sprintf("%s%s", label_prefix, 100 - 100 * as.numeric(available_iso_lvl))
   
   # prep hm color palette
   hm_color_palette_smooth <-
@@ -186,7 +186,6 @@ heatmap_combo_metrics <- function(
                        unique(dt_excess[get(cellline_name) == cl_name][[duration]]))
   # plot data
   dt_ <- dt_excess[, c(conc, conc_2, metric), with = FALSE]
-  
   
   x_axis_drug <- if (swap_axes) {
     drug1_name
@@ -465,14 +464,12 @@ heatmap_combo_metrics_panel <- function(
                         unique(dt_excess[get(cellline_name) == cl_name][[clid]]))
   # legend
   legend_title_iso <- "Iso Levels"
-  prefix <- if (normalization_type == "GR") {
+  label_prefix <- if (normalization_type == "GR") {
     "GR"
   } else {
     "IC"
   }
-  
-  legend_lbl_iso <- sprintf("%s%s", prefix, 100 - 100 * as.numeric(available_iso_lvl))
-  
+  legend_lbl_iso <- sprintf("%s%s", label_prefix, 100 - 100 * as.numeric(available_iso_lvl))
   
   # prep hm color palette
   hm_color_palette_smooth <-
@@ -500,17 +497,20 @@ heatmap_combo_metrics_panel <- function(
     
     dt_ <- dt_excess[, c(conc, conc_2, mx_name), with = FALSE]
     
-    x_axis_lab <- sprintf("%s [\U00B5M]", if (swap_axes) {
+    x_axis_drug <- if (swap_axes) {
       drug1_name
     } else {
       drug2_name
-    })
+    }
     
-    y_axis_lab <- sprintf("%s [\U00B5M]", if (swap_axes) {
+    y_axis_drug <- if (swap_axes) {
       drug2_name
     } else {
       drug1_name
-    })
+    }
+    
+    x_axis_lab <- sprintf("%s [\U00B5M]", x_axis_drug)
+    y_axis_lab <- sprintf("%s [\U00B5M]", y_axis_drug)
     
     if (!NROW(dt_) > 1 || # co-dilution input data is like: (conc = 0, conc_2 = 0, mx_name = 1)
         all(is.na(dt_[get(conc) != 0 & get(conc_2) != 0][[mx_name]]))) { # lack of smooth & excess data
@@ -834,14 +834,12 @@ plot_combination_index <- function(
   # check if isolines are available and adjust plotting logic accordingly
   if (NROW(available_iso_lvl) > 0) {
     legend_title_iso <- "Iso Levels"
-    prefix <- if (normalization_type == "GR") {
+    label_prefix <- if (normalization_type == "GR") {
       "GR"
     } else {
       "IC"
     }
-    
-    legend_lbl_iso <- sprintf("%s%s", prefix, 100 - 100 * as.numeric(available_iso_lvl))
-    
+    legend_lbl_iso <- sprintf("%s%s", label_prefix, 100 - 100 * as.numeric(available_iso_lvl))
     
     iso_colors <- 
       if (is.null(colors_vec_iso) || !all(vapply(colors_vec_iso, is_valid_color, logical(1)))) {
@@ -1052,17 +1050,21 @@ heatmap_combo_with_isoref <- function(
   # prep plot data
   dt_ <- dt_excess[, c(conc, conc_2, metric), with = FALSE]
   
-  x_axis_lab <- sprintf("%s [\U00B5M]", if (swap_axes) {
+  x_axis_drug <- if (swap_axes) {
     drug1_name
   } else {
     drug2_name
-  })
-  y_axis_lab <- sprintf("%s [\U00B5M]", if (swap_axes) {
+  }
+  
+  y_axis_drug <- if (swap_axes) {
     drug2_name
   } else {
     drug1_name
-  })
-
+  }
+  
+  x_axis_lab <- sprintf("%s [\U00B5M]", x_axis_drug)
+  y_axis_lab <- sprintf("%s [\U00B5M]", y_axis_drug)
+  
   if (!NROW(dt_) > 1) { # co-dilution input data is like: (conc = 0, conc_2 = 0, metric = 1)
     plt <- 
       ggplot2::ggplot() +
@@ -1151,13 +1153,13 @@ heatmap_combo_with_isoref <- function(
         iso_colors <- .get_iso_colors(available_iso_lvl)
         
         # plot
-        prefix <- if (normalization_type == "GR") {
+        label_prefix <- if (normalization_type == "GR") {
           "GR"
         } else {
           "IC"
         }
         
-        iso_label <- sprintf("%s%s", prefix, 100 - 100 * as.numeric(available_iso_lvl))
+        iso_label <- sprintf("%s%s", label_prefix, 100 - 100 * as.numeric(available_iso_lvl))
         names(iso_label) <- available_iso_lvl
         
         tab_measured <- dt_iso[, .SD, .SDcols = -c("pos_x_ref", "pos_y_ref")]
@@ -1411,17 +1413,20 @@ heatmap_combo_with_isoref_panel <- function(
                                gDRutils::prettify_flat_metrics(x = metric, human_readable = TRUE),
                                normalization_type)
   # base plot
-  x_axis_lab <- sprintf("%s [\U00B5M]", if (swap_axes) {
+  x_axis_drug <- if (swap_axes) {
     drug1_name
   } else {
     drug2_name
-  })
+  }
   
-  y_axis_lab <- sprintf("%s [\U00B5M]", if (swap_axes) {
+  y_axis_drug <- if (swap_axes) {
     drug2_name
   } else {
     drug1_name
-  })
+  }
+  
+  x_axis_lab <- sprintf("%s [\U00B5M]", x_axis_drug)
+  y_axis_lab <- sprintf("%s [\U00B5M]", y_axis_drug)
   
   plt <-
     ggplot2::ggplot(dt_tile,
@@ -1469,13 +1474,12 @@ heatmap_combo_with_isoref_panel <- function(
       iso_colors <- .get_iso_colors(available_iso_lvl)
       
       # plot
-      prefix <- if (normalization_type == "GR") {
+      label_prefix <- if (normalization_type == "GR") {
         "GR"
       } else {
         "IC"
       }
-      
-      iso_label <- sprintf("%s%s", prefix, 100 - 100 * as.numeric(iso_levels))
+      iso_label <- sprintf("%s%s", label_prefix, 100 - 100 * as.numeric(iso_levels))
       names(iso_label) <- iso_levels
       
       tab_measured <- dt_iso[, .SD, .SDcols = -c("pos_x_ref", "pos_y_ref")]